refactor(chat): Rename chat modules to kebab-case
src/lib/chat was the only module group using PascalCase
filenames for .ts and .svelte.ts modules; every other state
module in the repo is kebab-case. Rename the modules and
their test siblings, and update all imports. Svelte
components keep PascalCase per convention.

Resolves: #477
