+++ src/lib/chat/core/chat-core.svelte.ts
