+++ src/lib/chat/core/chat-draft-manager.svelte.ts
