+++ src/lib/chat/core/display-message-processor.ts
