+++ src/lib/chat/core/stream-processor.test.ts
