+++ src/lib/chat/ui/ChatMessage.svelte
@@ -6,7 +6,7 @@
 	import ChatReasoning from './ChatReasoning.svelte';
 	import MessageBubble from './MessageBubble.svelte';
 	import InlineEmailPrompt from './InlineEmailPrompt.svelte';
-	import { getChatUIContext } from './ChatContext.svelte.js';
+	import { getChatUIContext } from './chat-context.svelte.js';
 	import { deriveOrderedParts, LEADING_REASONING_KEY, type OrderedPart } from './ordered-parts.js';
 	import { type DisplayMessage, type Attachment } from '../core/types.js';
